refactor(sources): extract mode helpers
