+++ beamz/devices/sources/mode.py
@@ -1,1238 +1,51 @@
-import logging
-
 import jax.numpy as jnp
 import numpy as np
 
 from beamz.const import EPS_0, LIGHT_SPEED, MU_0
+from beamz.devices.sources.inject import (
+    _get_3d_huygens_terms,
+    _inject_3d_e_fields,
+    _inject_3d_fields,
+    _inject_3d_h_fields,
+    _inject_e_component,
+    _inject_h_component,
+    _match_shape,
+)
+from beamz.devices.sources.profiles import (
+    _axis_index_from_component_indices,
+    _backward_3d_mode_from_forward,
+    _build_3d_profiles,
+    _component_axis_coord,
+    _dominant_3d_pair,
+    _impedance_match_3d_tangential_pairs,
+    _impedance_match_e_profile,
+    _make_3d_mode_basis_profiles,
+    _modal_overlap_3d_profiles,
+    _modal_power_2d,
+    _modal_power_3d_from_profiles,
+    _normalize_2d_pair_by_power,
+    _normalize_3d_profiles_by_flux,
+    _numeric_impedance_axis,
+    _numeric_phase_delay,
+    _parse_direction,
+    _project_3d_profiles_to_real,
+    _remap_3d_solver_components,
+    _select_3d_impedance_index,
+    _select_3d_phase_ref,
+    _select_core_confined_mode_index,
+    _solve_numeric_k_axis,
+    _to_real_profile,
+)
 from beamz.devices.sources.solve import solve_modes
-
-logger = logging.getLogger(__name__)
-
-_VALID_DIRECTIONS = {"+x", "-x", "+y", "-y", "+z", "-z"}
-
-
-# ---------------------------------------------------------------------------
-# Private helpers
-# ---------------------------------------------------------------------------
-
-
-def _jax_tukey_window(M: int, alpha: float = 0.5) -> jnp.ndarray:
-    """JAX-compatible Tukey (tapered cosine) window.
-
-    Replaces scipy.signal.windows.tukey for differentiable source initialization.
-
-    Args:
-        M: Number of points in the window
-        alpha: Shape parameter (0 = rectangular, 1 = Hann)
-
-    Returns:
-        The Tukey window as a JAX array
-    """
-    if M <= 0:
-        return jnp.array([])
-    if M == 1:
-        return jnp.ones(1)
-
-    n = jnp.arange(M)
-    width = alpha * (M - 1) / 2.0
-
-    # Avoid division by zero when alpha=0
-    width = jnp.maximum(width, 1e-10)
-
-    # Three regions: taper up, flat, taper down
-    left_taper = 0.5 * (1 + jnp.cos(jnp.pi * (n / width - 1)))
-    right_taper = 0.5 * (1 + jnp.cos(jnp.pi * ((n - (M - 1 - width)) / width)))
-
-    window = jnp.where(
-        n < width, left_taper, jnp.where(n > (M - 1) - width, right_taper, 1.0)
-    )
-    return window
-
-
-def _scipy_tukey(n, alpha=0.3):
-    from scipy.signal.windows import tukey
-
-    return tukey(n, alpha=alpha)
-
-
-def _crop_and_window_2d(profile, z_s, z_e, t_s, t_e, window):
-    """Crop a 2D profile to [z_s:z_e, t_s:t_e] and multiply by window.
-
-    Handles shape mismatches by taking the minimum overlap region.
-    """
-    cropped = profile[z_s:z_e, t_s:t_e]
-    if cropped.size == 0:
-        return cropped
-    if cropped.shape == window.shape:
-        return cropped * window
-    z_min = min(cropped.shape[0], window.shape[0])
-    t_min = min(cropped.shape[1], window.shape[1])
-    return cropped[:z_min, :t_min] * window[:z_min, :t_min]
-
-
-def _make_tukey_window_2d(height_cells, width_cells, alpha=0.3, use_jax=True):
-    """Create a 2D Tukey window via outer product of 1D windows."""
-    _make = _jax_tukey_window if use_jax else _scipy_tukey
-    _ones = jnp.ones if use_jax else np.ones
-
-    wz = (
-        _make(height_cells, alpha=alpha)
-        if height_cells > 2
-        else _ones(max(1, height_cells))
-    )
-    wt = (
-        _make(width_cells, alpha=alpha)
-        if width_cells > 2
-        else _ones(max(1, width_cells))
-    )
-
-    if use_jax:
-        return wz[:, jnp.newaxis] * wt[jnp.newaxis, :]
-    return wz[:, np.newaxis] * wt[np.newaxis, :]
-
-
-def _stagger_half(field, axis):
-    """Average adjacent cells along *axis* (0 or 1) for Yee half-grid staggering."""
-    if field.shape[axis] <= 1:
-        return field
-    if axis == 0:
-        return 0.5 * (field[:-1, :] + field[1:, :])
-    return 0.5 * (field[:, :-1] + field[:, 1:])
-
-
-def _stagger_both(field):
-    """Stagger along both axes (for longitudinal H that sits at half-grid in both)."""
-    out = field
-    if out.shape[1] > 1:
-        out = 0.5 * (out[:, :-1] + out[:, 1:])
-    if out.shape[0] > 1:
-        out = 0.5 * (out[:-1, :] + out[1:, :])
-    return out
-
-
-def _compute_transverse_bounds(center_val, extent, resolution, grid_max):
-    """Return (start_idx, end_idx) for the injection window along a transverse axis."""
-    center_idx = int(round(center_val / resolution))
-    half_idx = int(round((extent / 2) / resolution))
-    return max(0, center_idx - half_idx), min(grid_max, center_idx + half_idx)
-
-
-def _impedance_match_e_profile(e_profile, h_profile, z_target, eps=1e-12):
-    """Scale E profile so the least-squares modal impedance matches ``z_target``."""
-    e = np.asarray(e_profile, dtype=np.complex128).reshape(-1)
-    h = np.asarray(h_profile, dtype=np.complex128).reshape(-1)
-    n = int(min(e.size, h.size))
-    if n <= 0:
-        return e_profile
-    e = e[:n]
-    h = h[:n]
-    denom = np.sum(h * np.conjugate(h))
-    if abs(denom) <= eps:
-        return e_profile
-    z_est = np.sum(e * np.conjugate(h)) / denom
-    z_mag = float(abs(z_est))
-    if (not np.isfinite(z_mag)) or z_mag <= eps:
-        return e_profile
-    scale = float(abs(z_target)) / z_mag
-    return np.asarray(e_profile) * scale
-
-
-def _modal_power_2d(e_profile, h_profile, signed_flux_sign, dl):
-    """Return 2D modal power using the same convention as port extraction."""
-    e = np.asarray(e_profile, dtype=np.complex128).reshape(-1)
-    h = np.asarray(h_profile, dtype=np.complex128).reshape(-1)
-    n = int(min(e.size, h.size))
-    if n <= 0:
-        return 0.0
-    e = e[:n]
-    h = h[:n]
-    p = 0.5 * np.real(np.sum(float(signed_flux_sign) * e * np.conjugate(h)) * float(dl))
-    return float(p)
-
-
-def _normalize_2d_pair_by_power(h_profile, e_profile, signed_flux_sign, dl, eps=1e-30):
-    """Normalize a 2D Huygens pair so |modal power| equals 1."""
-    h = np.asarray(h_profile)
-    e = np.asarray(e_profile)
-    p = _modal_power_2d(
-        np.asarray(e, dtype=np.complex128),
-        np.asarray(h, dtype=np.complex128),
-        signed_flux_sign=signed_flux_sign,
-        dl=dl,
-    )
-    if np.isfinite(p) and abs(p) > eps:
-        scale = np.sqrt(1.0 / abs(p))
-        h = h * scale
-        e = e * scale
-    return h, e
-
-
-def _to_real_profile(profile, imag_ratio_warn=1e-2, eps=1e-30):
-    """Project profile to real-valued injection coefficients."""
-    arr = np.asarray(profile, dtype=np.complex128)
-    re = np.real(arr)
-    im = np.imag(arr)
-    re_peak = float(np.max(np.abs(re))) if re.size else 0.0
-    im_peak = float(np.max(np.abs(im))) if im.size else 0.0
-    if re_peak > eps and im_peak / re_peak > imag_ratio_warn:
-        logger.debug(
-            "Mode profile has non-negligible imaginary content before real projection: "
-            "imag/real peak ratio=%.3e",
-            im_peak / re_peak,
-        )
-    return re
-
-
-def _solve_numeric_k_axis(omega, dt, d_axis, neff, eps=1e-30):
-    """Solve 1D Yee dispersion for k at a fixed omega (normal incidence)."""
-    neff_r = max(float(np.real(neff)), eps)
-    d = max(float(d_axis), eps)
-    dt_r = max(float(dt), eps)
-    omega_r = float(omega)
-    k_phys = omega_r * neff_r / LIGHT_SPEED
-
-    S = LIGHT_SPEED * dt_r / (neff_r * d)
-    if (not np.isfinite(S)) or S <= eps:
-        return k_phys
-
-    rhs = np.sin(0.5 * omega_r * dt_r) / S
-    rhs = float(np.clip(rhs, -1.0, 1.0))
-    k_num = (2.0 / d) * np.arcsin(rhs)
-    if (not np.isfinite(k_num)) or k_num <= eps:
-        return k_phys
-    return float(k_num)
-
-
-def _numeric_phase_delay(omega, k_num, delta_s, eps=1e-30):
-    """Convert numerical phase advance into a time delay."""
-    omega_r = max(abs(float(omega)), eps)
-    # Keep the sign: launch direction depends on the signed E/H plane offset.
-    return float((float(k_num) * float(delta_s)) / omega_r)
-
-
-def _numeric_impedance_axis(omega, dt, d_axis, k_num, neff, mu_r=1.0, eps=1e-30):
-    """Discrete normal-incidence Yee impedance for one propagation axis."""
-    neff_r = max(float(np.real(neff)), eps)
-    eta0 = np.sqrt(MU_0 / EPS_0)
-
-    denom = np.sin(0.5 * float(k_num) * float(d_axis))
-    if abs(denom) <= eps or not np.isfinite(denom):
-        return float(eta0 / neff_r)
-
-    numer = (
-        float(mu_r)
-        * MU_0
-        * (float(d_axis) / max(float(dt), eps))
-        * np.sin(0.5 * float(omega) * float(dt))
-    )
-    z_num = abs(numer / denom)
-    if (not np.isfinite(z_num)) or z_num <= eps:
-        return float(eta0 / neff_r)
-    return float(z_num)
-
-
-def _axis_index_from_component_indices(indices, axis):
-    """Extract scalar axis index from a 3D component index tuple."""
-    if indices is None:
-        return None
-    axis_pos = {"x": 2, "y": 1, "z": 0}[axis]
-    val = indices[axis_pos]
-    if isinstance(val, slice):
-        return None
-    return int(val)
-
-
-def _component_axis_coord(component_name, axis_index, axis, dx, dy, dz):
-    """Yee-location coordinate along propagation axis for one component plane index."""
-    if axis_index is None:
-        return 0.0
-
-    d_axis = {"x": dx, "y": dy, "z": dz}[axis]
-    staggered_along_axis = {
-        "x": {"Ex", "Hy", "Hz"},
-        "y": {"Ey", "Hx", "Hz"},
-        "z": {"Ez", "Hx", "Hy"},
-    }
-    offset = 1.0 if component_name in staggered_along_axis[axis] else 0.5
-    return (axis_index + offset) * d_axis
-
-
-def _dominant_3d_pair(axis, pol):
-    """Dominant launch pair used for 3D delay extraction."""
-    pair = {
-        ("x", "tm"): ("Ez", "Hy"),
-        ("x", "te"): ("Ey", "Hz"),
-        ("y", "tm"): ("Ez", "Hx"),
-        ("y", "te"): ("Ex", "Hz"),
-        ("z", "tm"): ("Ey", "Hx"),
-        ("z", "te"): ("Ex", "Hy"),
-    }.get((axis, pol))
-    if pair is None:
-        raise ValueError(f"Unsupported 3D delay pair for axis={axis!r}, pol={pol!r}")
-    return pair
-
-
-def _weighted_local_index(eps_profile_2d, components, eps=1e-30):
-    """Field-energy-weighted refractive index over a 2D cross-section."""
-    eps_arr = np.asarray(eps_profile_2d)
-    if eps_arr.size == 0:
-        return 1.0
-
-    w_sum = 0.0
-    ew_sum = 0.0
-    for comp in components:
-        arr = np.asarray(comp)
-        if arr.ndim != 2 or arr.size == 0:
-            continue
-        z = min(eps_arr.shape[0], arr.shape[0])
-        t = min(eps_arr.shape[1], arr.shape[1])
-        if z <= 0 or t <= 0:
-            continue
-        sl_eps = eps_arr[:z, :t]
-        sl_arr = arr[:z, :t]
-        w = np.abs(sl_arr) ** 2
-        w_sum += float(np.sum(w))
-        ew_sum += float(np.sum(sl_eps * w))
-
-    if w_sum <= eps:
-        return float(np.sqrt(max(float(np.mean(eps_arr)), eps)))
-    return float(np.sqrt(max(ew_sum / w_sum, eps)))
-
-
-def _select_3d_impedance_index(axis, pol, eps_profile_2d, Ex, Ey, Ez, Hx, Hy, Hz):
-    """Pick a local modal index for 3D impedance targeting."""
-    if pol == "tm":
-        tangential = {
-            "x": (Hy, Hz),
-            "y": (Hx, Hz),
-            "z": (Hx, Hy),
-        }[axis]
-    else:
-        tangential = {
-            "x": (Ey, Ez),
-            "y": (Ex, Ez),
-            "z": (Ex, Ey),
-        }[axis]
-    return _weighted_local_index(eps_profile_2d, tangential)
-
-
-def _impedance_match_3d_tangential_pairs(
-    axis, Ex_s, Ey_s, Ez_s, Hx_s, Hy_s, Hz_s, Z_target, eps=1e-12
-):
-    """Match tangential 3D E components to paired H components via LS impedance fit."""
-    e_map = {"Ex": Ex_s, "Ey": Ey_s, "Ez": Ez_s}
-    h_map = {"Hx": Hx_s, "Hy": Hy_s, "Hz": Hz_s}
-    pair_map = {
-        "x": [("Ey", "Hz"), ("Ez", "Hy")],
-        "y": [("Ex", "Hz"), ("Ez", "Hx")],
-        "z": [("Ex", "Hy"), ("Ey", "Hx")],
-    }
-    for e_name, h_name in pair_map[axis]:
-        e_field = np.asarray(e_map[e_name])
-        h_field = np.asarray(h_map[h_name])
-        abs_e = np.abs(e_field)
-        abs_h = np.abs(h_field)
-
-        # Robust local impedance estimate: use pointwise |E|/|H| over strong-field H support.
-        h_peak = float(np.max(abs_h))
-        if h_peak > eps:
-            mask = abs_h > (0.05 * h_peak)
-            if np.any(mask):
-                local_ratio = abs_e[mask] / (abs_h[mask] + eps)
-                ratio = float(np.median(local_ratio))
-            else:
-                ratio = 0.0
-        else:
-            ratio = 0.0
-
-        if not np.isfinite(ratio) or ratio <= eps:
-            e_norm = float(np.sqrt(np.sum(abs_e**2) + eps))
-            h_norm = float(np.sqrt(np.sum(abs_h**2) + eps))
-            ratio = e_norm / (h_norm + eps)
-
-        if not np.isfinite(ratio) or ratio <= eps:
-            continue
-        z_abs = float(abs(Z_target))
-        scale = z_abs / ratio
-        max_scale = max(20.0, 8.0 * z_abs)
-        scale = float(np.clip(scale, 1.0 / max_scale, max_scale))
-        e_map[e_name] = e_field * scale
-    return (
-        e_map["Ex"],
-        e_map["Ey"],
-        e_map["Ez"],
-    )
-
-
-def _parse_direction(direction):
-    """Parse and validate direction string into (axis, sign)."""
-    direction = str(direction).lower()
-    if direction not in _VALID_DIRECTIONS:
-        valid = ", ".join(sorted(_VALID_DIRECTIONS))
-        raise ValueError(f"direction must be one of {{{valid}}}, got {direction!r}")
-    axis = direction[1]
-    dir_sign = 1.0 if direction.startswith("+") else -1.0
-    return direction, axis, dir_sign
-
-
-def _remap_3d_solver_components(Ex, Ey, Ez, Hx, Hy, Hz, axis):
-    """Remap solver x-basis components to requested global propagation axis."""
-    if axis == "x":
-        return Ex, Ey, Ez, Hx, Hy, Hz
-    if axis == "y":
-        # Use the same right-handed x'->+y, y'->-x, z'->+z basis as the
-        # monitor-side modal extraction. Without the sign flip, y-directed 3D
-        # launches carry a persistent opposite-going modal component.
-        return -Ey, Ex, Ez, -Hy, Hx, Hz
-    if axis == "z":
-        # Cyclic remap from x-basis -> z-basis.
-        return Ey, Ez, Ex, Hy, Hz, Hx
-    raise ValueError(f"Unsupported axis {axis!r} for 3D remap")
-
-
-def _select_3d_phase_ref(axis, pol, Ex, Ey, Ez, Hx, Hy, Hz):
-    """Select 3D phase-reference field using the same H-based convention as 2D."""
-    preferred = {
-        ("x", "tm"): Hy,
-        ("x", "te"): Hz,
-        ("y", "tm"): Hx,
-        ("y", "te"): Hz,
-        ("z", "tm"): Hx,
-        ("z", "te"): Hy,
-    }
-    tangential_h = {
-        "x": (Hy, Hz),
-        "y": (Hx, Hz),
-        "z": (Hx, Hy),
-    }
-    key = (axis, pol)
-    if key not in preferred:
-        raise ValueError(f"Unsupported 3D phase-reference mapping for {key!r}")
-
-    ref_field = preferred[key]
-    if float(jnp.max(jnp.abs(ref_field))) >= 1e-9:
-        return ref_field
-
-    # Fallback for near-cutoff/degenerate cases: pick strongest tangential H component.
-    candidates = tangential_h[axis]
-    strengths = [float(jnp.max(jnp.abs(comp))) for comp in candidates]
-    return candidates[int(np.argmax(strengths))]
-
-
-def _select_core_confined_mode_index(eps_profile, e_fields, neff_values):
-    """Choose the candidate mode with strongest confinement in the high-index core."""
-    if e_fields is None or len(e_fields) <= 1:
-        return 0
-
-    eps_arr = np.asarray(eps_profile)
-    if eps_arr.size == 0:
-        return 0
-
-    eps_flat = np.real(eps_arr).ravel()
-    eps_min = float(np.min(eps_flat))
-    eps_max = float(np.max(eps_flat))
-    if (not np.isfinite(eps_min)) or (not np.isfinite(eps_max)) or (eps_max <= eps_min):
-        return 0
-
-    # Robust core estimate that also works with subpixel-smoothed profiles.
-    core_mask = eps_flat >= (eps_min + 0.5 * (eps_max - eps_min))
-    if not np.any(core_mask):
-        return 0
-
-    # Estimate core center for additional center-confinement scoring.
-    core_coords = np.argwhere(core_mask.reshape(eps_arr.shape))
-    if core_coords.size > 0:
-        core_center = np.mean(core_coords, axis=0)
-    else:
-        core_center = 0.5 * (np.asarray(eps_arr.shape, dtype=float) - 1.0)
-
-    best_idx = 0
-    best_core_frac = -np.inf
-    best_center_frac = -np.inf
-    best_neff = -np.inf
-
-    for idx, field in enumerate(e_fields):
-        f = np.asarray(field)
-        if f.ndim < 2:
-            continue
-
-        e_plane = np.sum(np.abs(f) ** 2, axis=0)
-        e_mag = e_plane.ravel()
-        n = min(e_mag.size, core_mask.size)
-        if n <= 0:
-            continue
-        e_mag = e_mag[:n]
-        mask = core_mask[:n]
-
-        total = float(np.sum(e_mag))
-        if (not np.isfinite(total)) or total <= 1e-30:
-            continue
-
-        core_frac = float(np.sum(e_mag[mask]) / total)
-
-        # Secondary discriminator: prefer modes concentrated near core center.
-        e_mag_shape = e_plane.shape
-        center_mask = np.ones(e_mag_shape, dtype=bool)
-        idx_grids = np.indices(e_mag_shape)
-        for axis_i, n_axis in enumerate(e_mag_shape):
-            half_span = max(1.0, 0.25 * float(n_axis))
-            center_mask &= (
-                np.abs(idx_grids[axis_i] - float(core_center[axis_i])) <= half_span
-            )
-        center_num = float(np.sum(e_plane[center_mask]))
-        center_frac = center_num / total
-        neff_r = float(np.real(neff_values[idx])) if idx < len(neff_values) else -np.inf
-
-        if (
-            (core_frac > best_core_frac + 1e-12)
-            or (
-                abs(core_frac - best_core_frac) <= 1e-12
-                and center_frac > best_center_frac + 1e-12
-            )
-            or (
-                abs(core_frac - best_core_frac) <= 1e-12
-                and abs(center_frac - best_center_frac) <= 1e-12
-                and neff_r > best_neff
-            )
-        ):
-            best_idx = idx
-            best_core_frac = core_frac
-            best_center_frac = center_frac
-            best_neff = neff_r
-
-    return int(best_idx)
-
-
-def _build_3d_profiles(
-    Ex,
-    Ey,
-    Ez,
-    Hx,
-    Hy,
-    Hz,
-    axis,
-    direction,
-    center,
-    width,
-    height,
-    center_idx,
-    offset_idx,
-    grid_shape,
-    resolution,
-    impedance_neff,
-    omega,
-    dt,
-):
-    """Build staggered, windowed, impedance-corrected injection profiles for 3D.
-
-    Returns
-    -------
-    profiles : dict
-        Mapping component name -> numpy real profile array.
-    indices : dict
-        Mapping component name -> index tuple for field injection.
-    """
-    nz, ny, nx = grid_shape
-    # Apply direction sign once while building source profiles.
-    dir_sign = 1.0 if direction.startswith("+") else -1.0
-    ETA_0 = np.sqrt(MU_0 / EPS_0)
-    neff_imp_r = max(float(np.real(impedance_neff)), 1e-6)
-    d_axis = resolution
-    if dt is not None:
-        k_num_imp = _solve_numeric_k_axis(omega, dt, d_axis, neff_imp_r)
-        Z_target = _numeric_impedance_axis(omega, dt, d_axis, k_num_imp, neff_imp_r)
-    else:
-        Z_target = ETA_0 / neff_imp_r
-
-    if axis == "x":
-        return _build_3d_x(
-            Ex,
-            Ey,
-            Ez,
-            Hx,
-            Hy,
-            Hz,
-            dir_sign,
-            Z_target,
-            center,
-            width,
-            height,
-            center_idx,
-            offset_idx,
-            nz,
-            ny,
-            nx,
-            resolution,
-        )
-    if axis == "y":
-        return _build_3d_y(
-            Ex,
-            Ey,
-            Ez,
-            Hx,
-            Hy,
-            Hz,
-            dir_sign,
-            Z_target,
-            center,
-            width,
-            height,
-            center_idx,
-            offset_idx,
-            nz,
-            ny,
-            nx,
-            resolution,
-        )
-    if axis == "z":
-        return _build_3d_z(
-            Ex,
-            Ey,
-            Ez,
-            Hx,
-            Hy,
-            Hz,
-            dir_sign,
-            Z_target,
-            center,
-            width,
-            height,
-            center_idx,
-            offset_idx,
-            nz,
-            ny,
-            nx,
-            resolution,
-        )
-    raise ValueError(f"Unsupported axis {axis!r} for 3D profile setup")
-
-
-def _build_3d_x(
-    Ex,
-    Ey,
-    Ez,
-    Hx,
-    Hy,
-    Hz,
-    dir_sign,
-    Z_target,
-    center,
-    width,
-    height,
-    center_idx,
-    offset_idx,
-    nz,
-    ny,
-    nx,
-    resolution,
-):
-    # --- stagger ---
-    Ey_s = _stagger_half(Ey, axis=1)
-    Ez_s = _stagger_half(Ez, axis=0)
-    Ex_s = Ex
-    Hx_s = _stagger_both(Hx)
-    Hy_s = _stagger_half(Hy, axis=0)
-    Hz_s = _stagger_half(Hz, axis=1)
-
-    # --- transverse bounds ---
-    y_start, y_end = _compute_transverse_bounds(center[1], width, resolution, ny)
-    z_center = center[2] if len(center) > 2 else (nz // 2) * resolution
-    z_start, z_end = _compute_transverse_bounds(z_center, height, resolution, nz)
-
-    # --- indices  (z_slice, y_slice, x_index) ---
-    indices = {
-        "Ex": (
-            slice(z_start, min(z_end, Ex_s.shape[0], nz)),
-            slice(y_start, min(y_end, Ex_s.shape[1], ny)),
-            offset_idx,
-        ),
-        "Ey": (
-            slice(z_start, min(z_end, Ey_s.shape[0], nz)),
-            slice(y_start, min(y_end, Ey_s.shape[1], ny - 1)),
-            center_idx,
-        ),
-        "Ez": (
-            slice(z_start, min(z_end, Ez_s.shape[0], nz - 1)),
-            slice(y_start, min(y_end, Ez_s.shape[1], ny)),
-            center_idx,
-        ),
-        "Hx": (
-            slice(z_start, min(z_end, Hx_s.shape[0], nz - 1)),
-            slice(y_start, min(y_end, Hx_s.shape[1], ny - 1)),
-            center_idx,
-        ),
-        "Hy": (
-            slice(z_start, min(z_end, Hy_s.shape[0], nz - 1)),
-            slice(y_start, min(y_end, Hy_s.shape[1], ny)),
-            offset_idx,
-        ),
-        "Hz": (
-            slice(z_start, min(z_end, Hz_s.shape[0], nz)),
-            slice(y_start, min(y_end, Hz_s.shape[1], ny - 1)),
-            offset_idx,
-        ),
-    }
-
-    # Match J/M balance using pairwise tangential impedance fits before windowing.
-    Ex_s, Ey_s, Ez_s = _impedance_match_3d_tangential_pairs(
-        "x",
-        Ex_s,
-        Ey_s,
-        Ez_s,
-        Hx_s,
-        Hy_s,
-        Hz_s,
-        Z_target,
-    )
-
-    # --- crop & window ---
-    staggered = {"Ex": Ex_s, "Ey": Ey_s, "Ez": Ez_s, "Hx": Hx_s, "Hy": Hy_s, "Hz": Hz_s}
-    profiles = _crop_and_window_all(
-        staggered,
-        z_start,
-        z_end,
-        y_start,
-        y_end,
-        dir_sign,
-        use_jax=True,
-        alpha=0.2,
-    )
-    d_area = float(resolution * resolution)
-    profiles = _normalize_3d_profiles_by_flux(profiles, axis="x", d_area=d_area)
-
-    extra = {
-        "_y_start": y_start,
-        "_y_end": y_end,
-        "_z_start": z_start,
-        "_z_end": z_end,
-        "_h_component": "Hy",
-        "_e_component": "Ey",
-    }
-    return profiles, indices, extra
-
-
-def _build_3d_y(
-    Ex,
-    Ey,
-    Ez,
-    Hx,
-    Hy,
-    Hz,
-    dir_sign,
-    Z_target,
-    center,
-    width,
-    height,
-    center_idx,
-    offset_idx,
-    nz,
-    ny,
-    nx,
-    resolution,
-):
-    # --- stagger (y-propagation) ---
-    Ex_s = _stagger_half(Ex, axis=1)
-    Ey_s = Ey
-    Ez_s = _stagger_half(Ez, axis=0)
-    Hx_s = _stagger_half(Hx, axis=0)
-    Hy_s = _stagger_both(Hy)
-    Hz_s = _stagger_half(Hz, axis=1)
-
-    # --- transverse bounds ---
-    x_start, x_end = _compute_transverse_bounds(center[0], width, resolution, nx)
-    z_center = center[2] if len(center) > 2 else (nz // 2) * resolution
-    z_start, z_end = _compute_transverse_bounds(z_center, height, resolution, nz)
-
-    # --- indices  (z_slice, y_index, x_slice) ---
-    indices = {
-        "Ex": (
-            slice(z_start, min(z_end, Ex_s.shape[0], nz)),
-            center_idx,
-            slice(x_start, min(x_end, Ex_s.shape[1], nx - 1)),
-        ),
-        "Ey": (
-            slice(z_start, min(z_end, Ey_s.shape[0], nz)),
-            offset_idx,
-            slice(x_start, min(x_end, Ey_s.shape[1], nx)),
-        ),
-        "Ez": (
-            slice(z_start, min(z_end, Ez_s.shape[0], nz - 1)),
-            center_idx,
-            slice(x_start, min(x_end, Ez_s.shape[1], nx)),
-        ),
-        "Hx": (
-            slice(z_start, min(z_end, Hx_s.shape[0], nz - 1)),
-            offset_idx,
-            slice(x_start, min(x_end, Hx_s.shape[1], nx)),
-        ),
-        "Hy": (
-            slice(z_start, min(z_end, Hy_s.shape[0], nz - 1)),
-            center_idx,
-            slice(x_start, min(x_end, Hy_s.shape[1], nx - 1)),
-        ),
-        "Hz": (
-            slice(z_start, min(z_end, Hz_s.shape[0], nz)),
-            offset_idx,
-            slice(x_start, min(x_end, Hz_s.shape[1], nx - 1)),
-        ),
-    }
-
-    # Match J/M balance using pairwise tangential impedance fits before windowing.
-    Ex_s, Ey_s, Ez_s = _impedance_match_3d_tangential_pairs(
-        "y",
-        Ex_s,
-        Ey_s,
-        Ez_s,
-        Hx_s,
-        Hy_s,
-        Hz_s,
-        Z_target,
-    )
-
-    # --- crop & window ---
-    staggered = {"Ex": Ex_s, "Ey": Ey_s, "Ez": Ez_s, "Hx": Hx_s, "Hy": Hy_s, "Hz": Hz_s}
-    profiles = _crop_and_window_all(
-        staggered,
-        z_start,
-        z_end,
-        x_start,
-        x_end,
-        dir_sign,
-        use_jax=False,
-        alpha=0.2,
-    )
-    d_area = float(resolution * resolution)
-    profiles = _normalize_3d_profiles_by_flux(profiles, axis="y", d_area=d_area)
-
-    extra = {
-        "_x_start": x_start,
-        "_x_end": x_end,
-        "_z_start": z_start,
-        "_z_end": z_end,
-        "_h_component": "Hx",
-        "_e_component": "Ex",
-    }
-    return profiles, indices, extra
-
-
-def _build_3d_z(
-    Ex,
-    Ey,
-    Ez,
-    Hx,
-    Hy,
-    Hz,
-    dir_sign,
-    Z_target,
-    center,
-    width,
-    height,
-    center_idx,
-    offset_idx,
-    nz,
-    ny,
-    nx,
-    resolution,
-):
-    # --- stagger (z-propagation) ---
-    Ex_s = _stagger_half(Ex, axis=1)
-    Ey_s = _stagger_half(Ey, axis=0)
-    Ez_s = Ez
-    Hx_s = _stagger_half(Hx, axis=0)
-    Hy_s = _stagger_half(Hy, axis=1)
-    Hz_s = _stagger_both(Hz)
-
-    # --- transverse bounds ---
-    x_start, x_end = _compute_transverse_bounds(center[0], width, resolution, nx)
-    y_start, y_end = _compute_transverse_bounds(center[1], height, resolution, ny)
-
-    e_z_idx = int(np.clip(center_idx, 0, nz - 1))
-    h_z_idx = int(np.clip(offset_idx, 0, max(nz - 2, 0)))
-    ez_z_idx = int(np.clip(center_idx, 0, max(nz - 2, 0)))
-    hz_z_idx = int(np.clip(offset_idx, 0, nz - 1))
-
-    # --- indices  (z_index, y_slice, x_slice) ---
-    indices = {
-        "Ex": (
-            e_z_idx,
-            slice(y_start, min(y_end, Ex_s.shape[0], ny)),
-            slice(x_start, min(x_end, Ex_s.shape[1], nx - 1)),
-        ),
-        "Ey": (
-            e_z_idx,
-            slice(y_start, min(y_end, Ey_s.shape[0], ny - 1)),
-            slice(x_start, min(x_end, Ey_s.shape[1], nx)),
-        ),
-        "Ez": (
-            ez_z_idx,
-            slice(y_start, min(y_end, Ez_s.shape[0], ny)),
-            slice(x_start, min(x_end, Ez_s.shape[1], nx)),
-        ),
-        "Hx": (
-            h_z_idx,
-            slice(y_start, min(y_end, Hx_s.shape[0], ny - 1)),
-            slice(x_start, min(x_end, Hx_s.shape[1], nx)),
-        ),
-        "Hy": (
-            h_z_idx,
-            slice(y_start, min(y_end, Hy_s.shape[0], ny)),
-            slice(x_start, min(x_end, Hy_s.shape[1], nx - 1)),
-        ),
-        "Hz": (
-            hz_z_idx,
-            slice(y_start, min(y_end, Hz_s.shape[0], ny - 1)),
-            slice(x_start, min(x_end, Hz_s.shape[1], nx - 1)),
-        ),
-    }
-
-    # Match J/M balance using pairwise tangential impedance fits before windowing.
-    Ex_s, Ey_s, Ez_s = _impedance_match_3d_tangential_pairs(
-        "z",
-        Ex_s,
-        Ey_s,
-        Ez_s,
-        Hx_s,
-        Hy_s,
-        Hz_s,
-        Z_target,
-    )
-
-    # --- crop & window ---
-    staggered = {"Ex": Ex_s, "Ey": Ey_s, "Ez": Ez_s, "Hx": Hx_s, "Hy": Hy_s, "Hz": Hz_s}
-    profiles = _crop_and_window_all(
-        staggered,
-        y_start,
-        y_end,
-        x_start,
-        x_end,
-        dir_sign,
-        use_jax=True,
-        alpha=0.2,
-    )
-    d_area = float(resolution * resolution)
-    profiles = _normalize_3d_profiles_by_flux(profiles, axis="z", d_area=d_area)
-
-    extra = {
-        "_x_start": x_start,
-        "_x_end": x_end,
-        "_y_start": y_start,
-        "_y_end": y_end,
-        "_h_component": "Hx",
-        "_e_component": "Ex",
-    }
-    return profiles, indices, extra
-
-
-def _crop_and_window_all(
-    staggered,
-    z_start,
-    z_end,
-    t_start,
-    t_end,
-    dir_sign,
-    use_jax,
-    alpha=0.3,
-):
-    """Crop all six staggered profiles and multiply by a 2D Tukey window."""
-    ref = next(iter(staggered.values()))
-    pz_end = min(z_end, ref.shape[0])
-    pt_end = min(t_end, ref.shape[1])
-    h_cells = pz_end - z_start
-    w_cells = pt_end - t_start
-
-    window = _make_tukey_window_2d(h_cells, w_cells, alpha=alpha, use_jax=use_jax)
-
-    profiles = {}
-    for name, field in staggered.items():
-        fe = min(z_end, field.shape[0])
-        te = min(t_end, field.shape[1])
-        profiles[name] = dir_sign * np.real(
-            _crop_and_window_2d(field, z_start, fe, t_start, te, window)
-        )
-    return profiles
-
-
-def _modal_power_3d_from_profiles(profiles, axis, d_area):
-    """Compute 3D modal power from profiles on a cross-section."""
-    ex = profiles.get("Ex")
-    ey = profiles.get("Ey")
-    ez = profiles.get("Ez")
-    hx = profiles.get("Hx")
-    hy = profiles.get("Hy")
-    hz = profiles.get("Hz")
-    if any(v is None for v in (ex, ey, ez, hx, hy, hz)):
-        return 0.0
-
-    ex = np.asarray(ex, dtype=np.complex128)
-    ey = np.asarray(ey, dtype=np.complex128)
-    ez = np.asarray(ez, dtype=np.complex128)
-    hx = np.asarray(hx, dtype=np.complex128)
-    hy = np.asarray(hy, dtype=np.complex128)
-    hz = np.asarray(hz, dtype=np.complex128)
-    if ex.ndim == 1:
-        ex = ex[:, None]
-    if ey.ndim == 1:
-        ey = ey[:, None]
-    if ez.ndim == 1:
-        ez = ez[:, None]
-    if hx.ndim == 1:
-        hx = hx[:, None]
-    if hy.ndim == 1:
-        hy = hy[:, None]
-    if hz.ndim == 1:
-        hz = hz[:, None]
-    ny = min(
-        ex.shape[0], ey.shape[0], ez.shape[0], hx.shape[0], hy.shape[0], hz.shape[0]
-    )
-    nx = min(
-        ex.shape[1], ey.shape[1], ez.shape[1], hx.shape[1], hy.shape[1], hz.shape[1]
-    )
-    if ny <= 0 or nx <= 0:
-        return 0.0
-
-    ex = ex[:ny, :nx]
-    ey = ey[:ny, :nx]
-    ez = ez[:ny, :nx]
-    hx = hx[:ny, :nx]
-    hy = hy[:ny, :nx]
-    hz = hz[:ny, :nx]
-
-    if axis == "x":
-        s_axis = ey * np.conjugate(hz) - ez * np.conjugate(hy)
-    elif axis == "y":
-        s_axis = ez * np.conjugate(hx) - ex * np.conjugate(hz)
-    else:
-        s_axis = ex * np.conjugate(hy) - ey * np.conjugate(hx)
-    return float(0.5 * np.real(np.sum(s_axis) * float(d_area)))
-
-
-def _normalize_3d_profiles_by_flux(profiles, axis, d_area=1.0, eps=1e-18):
-    """Normalize 3D source profiles so |modal power| equals 1."""
-    flux = _modal_power_3d_from_profiles(profiles, axis=axis, d_area=d_area)
-    if (not np.isfinite(flux)) or abs(flux) <= eps:
-        return profiles
-
-    scale = float(np.sqrt(1.0 / max(abs(flux), eps)))
-    scale = float(np.clip(scale, 1e-6, 1e6))
-    for key, value in profiles.items():
-        profiles[key] = np.asarray(value) * scale
-    return profiles
-
-
-def _backward_3d_mode_from_forward(profiles):
-    """Return the backward-going counterpart of a forward 3D modal field set."""
-    out = {}
-    for key, value in profiles.items():
-        arr = np.asarray(value, dtype=np.complex128)
-        out[key] = -arr if key.startswith("H") else arr.copy()
-    return out
-
-
-def _make_3d_mode_basis_profiles(profiles, axis, d_area=1.0):
-    """Build unit-flux forward/backward 3D basis fields from one solved mode."""
-    forward = {
-        key: np.asarray(value, dtype=np.complex128) for key, value in profiles.items()
-    }
-    forward = _normalize_3d_profiles_by_flux(forward, axis=axis, d_area=d_area)
-    backward = _backward_3d_mode_from_forward(forward)
-    return forward, backward
-
-
-def _modal_overlap_3d_profiles(field_profiles, mode_profiles, axis, d_area):
-    """Symmetric power overlap between a field sample and a 3D modal basis field."""
-    comp_map = {
-        "x": ("Ey", "Ez", "Hz", "Hy"),
-        "y": ("Ez", "Ex", "Hx", "Hz"),
-        "z": ("Ex", "Ey", "Hy", "Hx"),
-    }
-    try:
-        e1, e2, h1, h2 = comp_map[str(axis)]
-    except KeyError as exc:
-        raise ValueError(f"Unsupported axis {axis!r}") from exc
-
-    arrays = {}
-    n_common = None
-    for name in (e1, e2, h1, h2):
-        f_arr = np.asarray(field_profiles[name], dtype=np.complex128).reshape(-1)
-        m_arr = np.asarray(mode_profiles[name], dtype=np.complex128).reshape(-1)
-        n_local = int(min(f_arr.size, m_arr.size))
-        if n_local <= 0:
-            return np.complex128(0.0 + 0.0j)
-        n_common = n_local if n_common is None else min(n_common, n_local)
-        arrays[name] = (f_arr, m_arr)
-
-    n_common = int(max(0, n_common or 0))
-    if n_common <= 0:
-        return np.complex128(0.0 + 0.0j)
-
-    ef1 = arrays[e1][0][:n_common]
-    ef2 = arrays[e2][0][:n_common]
-    hf1 = arrays[h1][0][:n_common]
-    hf2 = arrays[h2][0][:n_common]
-    em1 = arrays[e1][1][:n_common]
-    em2 = arrays[e2][1][:n_common]
-    hm1 = arrays[h1][1][:n_common]
-    hm2 = arrays[h2][1][:n_common]
-
-    overlap = (
-        0.25
-        * np.sum(
-            ef1 * np.conjugate(hm1)
-            - ef2 * np.conjugate(hm2)
-            + np.conjugate(em1) * hf1
-            - np.conjugate(em2) * hf2
-        )
-        * float(d_area)
-    )
-    return np.complex128(overlap)
-
-
-def _project_3d_profiles_to_real(profiles):
-    """Project 3D source profiles to real-valued runtime injection arrays."""
-    out = {}
-    for key, value in profiles.items():
-        out[key] = _to_real_profile(value)
-    return out
-
-
-# ---------------------------------------------------------------------------
-# Huygens cross-product sign tables
-# sign = multiplier in: target += sign * source_profile * sig * dt / (material * dx)
-# Only transverse components are injected; longitudinal components are skipped.
-# ---------------------------------------------------------------------------
-_HUYGENS_SIGNS = {
-    "x": {
-        "e": [("Ey", "Hz", -1), ("Ez", "Hy", +1)],
-        "h": [("Hy", "Ez", -1), ("Hz", "Ey", +1)],
-    },
-    "y": {
-        "e": [("Ex", "Hz", -1), ("Ez", "Hx", +1)],
-        "h": [("Hx", "Ez", +1), ("Hz", "Ex", -1)],
-    },
-    "z": {
-        "e": [("Ex", "Hy", -1), ("Ey", "Hx", +1)],
-        "h": [("Hx", "Ey", +1), ("Hy", "Ex", -1)],
-    },
-}
-
-
-def _get_3d_huygens_terms(axis, pol):
-    """Return 3D sign terms with TE gauge parity matched to 2D conventions."""
-    e_terms = list(_HUYGENS_SIGNS[axis]["e"])
-    h_terms = list(_HUYGENS_SIGNS[axis]["h"])
-    return e_terms, h_terms
-
-
-def _inject_3d_e_fields(fields, profiles, indices, signal_e, dt, resolution, axis, pol):
-    """Inject E-field components for 3D Huygens source (J = n x H), after E update."""
-    e_terms, _ = _get_3d_huygens_terms(axis, pol)
-    for e_comp, h_source, sign in e_terms:
-        _inject_e_component(
-            fields,
-            e_comp,
-            profiles,
-            indices,
-            h_source,
-            signal_e,
-            dt,
-            resolution,
-            sign=sign,
-        )
-
-
-def _inject_3d_h_fields(fields, profiles, indices, signal_h, dt, resolution, axis, pol):
-    """Inject H-field components for 3D Huygens source (M = -n x E), after H update."""
-    _, h_terms = _get_3d_huygens_terms(axis, pol)
-    for h_comp, e_source, sign in h_terms:
-        _inject_h_component(
-            fields,
-            h_comp,
-            profiles,
-            indices,
-            e_source,
-            signal_h,
-            dt,
-            resolution,
-            sign=sign,
-        )
-
-
-def _inject_3d_fields(
-    fields, profiles, indices, signal_e, signal_h, dt, resolution, axis="x", pol="tm"
-):
-    """Inject all field components into a 3D field object (backward compat wrapper)."""
-    _inject_3d_h_fields(fields, profiles, indices, signal_h, dt, resolution, axis, pol)
-    _inject_3d_e_fields(fields, profiles, indices, signal_e, dt, resolution, axis, pol)
-
-
-def _inject_e_component(
-    fields, comp, profiles, indices, j_source, sig, dt, res, sign=-1
-):
-    """Inject one E-field component via J = cross(n, H)."""
-    profile = profiles.get(comp)
-    idx = indices.get(comp)
-    j_term = profiles.get(j_source)
-    if profile is None or idx is None or j_term is None:
-        return
-    target = getattr(fields, comp)[idx]
-    j_term = _match_shape(j_term, target.shape)
-    if j_term is None:
-        logger.debug("Shape mismatch injecting %s, skipping", comp)
-        return
-    eps = fields.permittivity[idx]
-    setattr(
-        fields,
-        comp,
-        getattr(fields, comp)
-        .at[idx]
-        .add(sign * j_term * sig * dt / (EPS_0 * eps * res)),
-    )
-
-
-def _inject_h_component(
-    fields, comp, profiles, indices, m_source, sig, dt, res, sign=-1
-):
-    """Inject one H-field component via M = -cross(n, E)."""
-    profile = profiles.get(comp)
-    idx = indices.get(comp)
-    m_term = profiles.get(m_source)
-    if profile is None or idx is None or m_term is None:
-        return
-    target = getattr(fields, comp)[idx]
-    m_term = _match_shape(m_term, target.shape)
-    if m_term is None:
-        logger.debug("Shape mismatch injecting %s, skipping", comp)
-        return
-    mu = getattr(fields, "permeability", None)
-    mu_val = mu[idx] if mu is not None else 1.0
-    setattr(
-        fields,
-        comp,
-        getattr(fields, comp)
-        .at[idx]
-        .add(sign * m_term * sig * dt / (MU_0 * mu_val * res)),
-    )
-
-
-def _match_shape(profile, target_shape):
-    """Match profile shape to target field shape, trimming or padding as needed."""
-    if profile is None:
-        return None
-    profile = np.squeeze(profile)
-    if profile.shape == target_shape:
-        return profile
-    if profile.ndim == len(target_shape):
-        slices = tuple(
-            slice(0, min(profile.shape[i], target_shape[i]))
-            for i in range(profile.ndim)
-        )
-        trimmed = profile[slices]
-        if trimmed.shape == target_shape:
-            return trimmed
-        result = np.zeros(target_shape, dtype=profile.dtype)
-        result[tuple(slice(0, trimmed.shape[i]) for i in range(trimmed.ndim))] = trimmed
-        return result
-    return None
+from beamz.devices.sources.windows import (
+    _compute_transverse_bounds,
+    _crop_and_window_2d,
+    _jax_tukey_window,
+    _make_tukey_window_2d,
+    _scipy_tukey,
+    _stagger_both,
+    _stagger_half,
+)
 
 
 # ---------------------------------------------------------------------------
